:recycle:[refactor]: 순환참조 에러 해결
